integrate staff pics
